requirements.txt
asgiref==3.2.3
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==1.6.1
Markdown==3.2.2
mysqlclient==1.4.6
Pillow==7.0.0
PyJWT==1.7.1
python-dotenv==0.11.0
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.1.0
zipp==3.1.0
